extract predictions components
