+++ src/components/route-browser/index.js
@@ -1,6 +1,6 @@
 export * from './context'
 export * from './navigation'
-export * from './predictions-graph'
-export * from './predictions-list'
+export * from './route-predictions'
 export * from './scene-metadata'
+export * from './scene-predictions'
 export * from './scene-prefetch'
