+++ src/components/route-browser/scene-predictions/index.js
@@ -1,0 +1,1 @@
+export * from './list'
